requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
asyncpg==0.29.0
influxdb-client==1.46.0
aiokafka==0.11.0
paho-mqtt==2.1.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.0
pymodbus==3.8.2
asyncua==1.1.5
python-multipart==0.0.12
email-validator==2.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
